Make exporter module private
